% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_function.R
\name{cbind.fill}
\alias{cbind.fill}
\title{The functions in this file contain other additional functions to support DMU
Cbind function source: http://stackoverflow.com/questions/7962267/cbind-a-df-with-an-empty-df-cbind-fill}
\usage{
\method{cbind}{fill}(...)
}
\description{
The functions in this file contain other additional functions to support DMU
Cbind function source: http://stackoverflow.com/questions/7962267/cbind-a-df-with-an-empty-df-cbind-fill
}
