% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMU_hyperpara_optimisation.R
\name{hyperpara_optimize}
\alias{hyperpara_optimize}
\title{The function in this file perform hyperparameter optimisation for DMU.}
\usage{
hyperpara_optimize(optimtype = "opt_GA", traindata = NA, testdata = NA, ...)
}
\description{
The function in this file perform hyperparameter optimisation for DMU.
}
