% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artificial_datagen.R
\name{randcorr}
\alias{randcorr}
\title{The functions in this file generate the artificial dataset for the DMU modeling}
\usage{
randcorr(varnum = 5, maxcorr = 0.5, type = c("upperlist", "matrix"), seed = 1)
}
\description{
The functions in this file generate the artificial dataset for the DMU modeling
}
