% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standard_techniques.R
\name{reg_results}
\alias{reg_results}
\title{The funcstions in this file focus determining the performance of the models using linear regression and predictive mean matching based MICE.}
\usage{
reg_results(hyperpara, data_prep)
}
\description{
The funcstions in this file focus determining the performance of the models using linear regression and predictive mean matching based MICE.
}
