% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_exeution.R
\name{runmodel}
\alias{runmodel}
\title{The functions in this file prepare the models}
\usage{
runmodel(
  technique,
  traindata,
  testdata,
  trainset = datatype,
  seed = seed_run,
  cluster_number = sample_clustersize,
  ...
)
}
\arguments{
\item{seed}{is for result repetition.}

\item{miss}{is the maximum percentage of missing values allowed in the feature. All features with more missing values will be dropped from the analysis process.}

\item{correlation}{is the maximum positive and negative correlation allowed in the predictor correlation matrix. All features with higher correlation values are dropped from the analysis process.}

\item{dataset}{is the code for the dataset for which this study will be performed. 1: is for CHSI dataset and 3: is for SWAN dataset.}

\item{sample_clustersize}{is the number of pieces the samples has to be fragmented for DMU.}

\item{sample_to_featureratio}{is the minimum sample to feature ratio allowed in each of the cluster. Any cluster with less than the minimum sample to feature ratio is dropped from the analysis process.}

\item{datatype}{determines whether the training data will have any sample with complete data or not. "Both" means that training data will have samples containing complete and incomplete data. "Incomplete" means that training data will have samples containing only incomplete data.}
}
\description{
The functions in this file prepare the models
}
