% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Performance.R
\name{y_est_nointerac}
\alias{y_est_nointerac}
\title{The funcstions in this file focus determining the performance of the models}
\usage{
y_est_nointerac(betadf, xtestdf, meanbeta)
}
\description{
The funcstions in this file focus determining the performance of the models
}
